refactor: simplify x/nft dependencies
