+++ tests/integration/nft/app_config.go
@@ -1,4 +1,4 @@
-package testutil
+package nft
 
 import (
 	"cosmossdk.io/core/appconfig"
