+++ tests/integration/nft/client/cli/query_test.go
